Revamp to remove Python constraint
As Ninja is a build tool with no real Python dependency constraint, we
don't want it to unnecessarily be tied to some VC runtime environment.
Instead we want to allow it to install in a variety of environments
including alongside Python 2.7 (if needed). So we sidestep this
constraint by asking `conda-build` to ignore `vc` for the `build` and
merely require the expected VC runtime.